Dev compression

+++ include/codage.h
@@ -13,7 +13,7 @@ phtree_t arbre_canonique(uint8_t prof[256]);
 /*
  *
  */
-void correspondance(phtree_t t, uint8_t code[256]);
+void correspondance(phtree_t t, uint64_t code[256]);
 
 
 
@@ -36,7 +36,7 @@ phtree_t get_file(file_fifo* f);
 int isEmpty(file_fifo* f);
 void ecriture_tableau( char racine[256]);
 void delete_null_prof(uint8_t prof[256], char racine[256]);
-void init_tab( uint8_t tab[256]);
+void init_tab( uint64_t tab[256]);
 void tri_tableau_racine_etape1(uint8_t prof[256], char racine[256]);
 void tri_tableau_racine_etape2(uint8_t prof[256], char racine[256]);
 void tri_tableau_index(char racine[256], int debut, int fin);
