+++ include/io.h
@@ -1,6 +1,6 @@
 #include <stdint.h>
 
 
-void transcodage(char* file_in, char* file_out, uint8_t code[256], uint8_t prof[256]);
+void transcodage(char* file_in, char* file_out, uint64_t code[256], uint8_t prof[256]);
 
 void decodage(char* file_in, char* file_out);
